requirements.common.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asyncpg==0.30.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
Brotli==1.1.0
browser-cookie3==0.20.1
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi[all]==0.115.6
fastapi-cli==0.0.7
filelock==3.17.0
fonttools==4.56.0
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
instaloader==4.14.1
jeepney==0.8.0
Jinja2==3.1.5
kiwisolver==1.4.8
lxml==5.3.1
lz4==4.4.3
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mutagen==1.47.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycryptodomex==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytubefix==8.12.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
seaborn==0.13.2
shadowcopy==0.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.28
starlette==0.41.3
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.1
WMI==1.4.9
yt-dlp==2025.2.19
redis==5.2.1
IMAPClient==3.0.1